+++ packages/browser-wallet/src/popup/pages/Account/Account.tsx
@@ -1,16 +1,24 @@
+import clsx from 'clsx';
 import { useAtom, useAtomValue } from 'jotai';
-import React, { useCallback } from 'react';
+import React, { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { Outlet, Route, Routes, useLocation, useNavigate } from 'react-router-dom';
+
 import { accountsAtom, selectedAccountAtom } from '@popup/store/account';
 import Button from '@popup/shared/Button';
 import { credentialsAtom, urlWhitelistAtom } from '@popup/store/settings';
+import CloseButton from '@popup/shared/CloseButton';
+import { absoluteRoutes } from '@popup/constants/routes';
+import MenuButton from '@popup/shared/MenuButton';
+import { accountRoutes } from './routes';
+import AccountActions from './AccountActions';
+import DisplayAddress from './DisplayAddress';
 import TransactionList from './TransactionList';
 
-export default function Account() {
+function AccountSettings() {
     const { t } = useTranslation('account');
-    const accounts = useAtomValue(accountsAtom);
-    const [selectedAccount, setSelectedAccount] = useAtom(selectedAccountAtom);
     const [creds, setCreds] = useAtom(credentialsAtom);
+    const [selectedAccount, setSelectedAccount] = useAtom(selectedAccountAtom);
     const [whitelist, setWhitelist] = useAtom(urlWhitelistAtom);
 
     const removeAccount = useCallback(() => {
@@ -24,29 +32,77 @@ export default function Account() {
         setWhitelist([]);
     }, []);
 
+    if (selectedAccount === undefined) {
+        return null;
+    }
+
     return (
-        <>
-            <div className="flex-column justify-center align-center">
-                <div className="flex justify-space-between w-full">
-                    {accounts.length === 0 && <div>{t('noAccounts')}</div>}
-                </div>
+        <div className="flex-column">
+            <Button danger onClick={removeAccount}>
+                {t('removeAccount')}
+            </Button>
+            <Button disabled={!whitelist.length} danger className="m-t-20" onClick={removeConnections}>
+                {t('resetConnections')}
+            </Button>
+        </div>
+    );
+}
+
+function Account() {
+    const { t } = useTranslation('account');
+    const accounts = useAtomValue(accountsAtom);
+    const selectedAccount = useAtomValue(selectedAccountAtom);
+    const { pathname } = useLocation();
+    const nav = useNavigate();
+    const [detailsExpanded, setDetailsExpanded] = useState(true);
+
+    const canClose = pathname !== absoluteRoutes.home.account.path;
+
+    return (
+        <div className="flex-column justify-space-between align-center h-full relative">
+            <MenuButton
+                className="account-page__hide"
+                open={detailsExpanded}
+                onClick={() => setDetailsExpanded((o) => !o)}
+            />
+            <div className="account-page__content">
+                {accounts.length === 0 && <div>{t('noAccounts')}</div>}
                 {selectedAccount !== undefined && (
                     <>
-                        <div className="account-page__address">{t('address', { address: selectedAccount })}</div>
-                        <Button danger className="m-t-20" onClick={removeAccount}>
-                            {t('removeAccount')}
-                        </Button>
+                        <div
+                            className={clsx(
+                                'account-page__details',
+                                detailsExpanded && 'account-page__details--expanded'
+                            )}
+                        >
+                            {selectedAccount}
+                        </div>
+                        <div className="account-page__routes">
+                            <Outlet />
+                            {canClose && (
+                                <CloseButton
+                                    className="account-page__close"
+                                    onClick={() => nav(absoluteRoutes.home.account.path)}
+                                />
+                            )}
+                        </div>
                     </>
                 )}
-                <Button disabled={!whitelist.length} danger className="m-b-20 m-t-10" onClick={removeConnections}>
-                    {t('resetConnections')}
-                </Button>
             </div>
-            {selectedAccount !== undefined && (
-                <div className="account-page__transaction-list">
-                    <TransactionList key={selectedAccount} accountAddress={selectedAccount} />
-                </div>
-            )}
-        </>
+            <AccountActions className="account-page__actions" />
+        </div>
+    );
+}
+
+export default function AccountRoutes() {
+    return (
+        <Routes>
+            <Route element={<Account />}>
+                <Route index element={<TransactionList />} />
+                <Route path={accountRoutes.send} element={<div>Send CCD</div>} />
+                <Route path={accountRoutes.receive} element={<DisplayAddress />} />
+                <Route path={accountRoutes.settings} element={<AccountSettings />} />
+            </Route>
+        </Routes>
     );
 }