+++ packages/browser-wallet/src/popup/pages/Account/i18n/da.ts
@@ -1,10 +1,16 @@
 import type en from './en';
 
 const t: typeof en = {
-    address: 'Adresse: {{address}}',
     noAccounts: 'Ingen konti i wallet',
     removeAccount: 'Fjern konto (kun lokalt)',
     resetConnections: 'Fjern forbindelser',
+    accountAddress: 'Konto adresse',
+    actions: {
+        log: 'transaktionslog',
+        send: 'send ccd',
+        receive: 'modtag ccd',
+        settings: 'kontoindstillinger',
+    },
 };
 
 export default t;