+++ packages/browser-wallet/src/popup/pages/Account/i18n/en.ts
@@ -1,8 +1,14 @@
 const t = {
-    address: 'Address: {{address}}',
     noAccounts: 'No accounts in wallet',
     removeAccount: 'Remove account (local only)',
     resetConnections: 'Reset connections',
+    accountAddress: 'Account address',
+    actions: {
+        log: 'transaction log',
+        send: 'send ccd',
+        receive: 'receive ccd',
+        settings: 'account settings',
+    },
 };
 
 export default t;