+++ packages/browser-wallet/CHANGELOG.md
@@ -4,12 +4,19 @@
 
 ### Added
 
+-   Added display of address as text and qr code in the "receive ccd" section.
+
+## 0.2.0
+
+### Added
+
 -   Signing intializeSmartContract transactions.
 -   Added menu for navigating between accounts, id cards, and settings (accessible by clicking Concordium logo).
 -   Settings page to control individual settings, including visual settings (dark/light mode), network settings along with placeholder pages for settings to come at a later stage.
 -   Add single credential functionality accessible from account page
 -   Remove account (from local storage) from account page
 -   An event is emitted when the JSON-RPC configuration is updated.
+-   Navigation on the bottom of the account page to navigate between sections of the page.
 
 ### Changed